CTF workflow refactored
